+++ js/RTCPeerConnection.js
@@ -18,7 +18,6 @@ var debug = require('debug')('iosrtc:RTCPeerConnection'),
 	RTCRtpReceiver = require('./RTCRtpReceiver'),
 	RTCRtpSender = require('./RTCRtpSender'),
 	{ RTCRtpTransceiver, addTransceiverToPeerConnection } = require('./RTCRtpTransceiver'),
-	RTCStatsResponse = require('./RTCStatsResponse'),
 	RTCStatsReport = require('./RTCStatsReport'),
 	MediaStream = require('./MediaStream'),
 	{ MediaStreamTrack, newMediaStreamTrackId } = require('./MediaStreamTrack'),
@@ -55,8 +54,16 @@ function RTCPeerConnection(pcConfig, pcConstraints) {
 		'getLocalStreams',
 		RTCPeerConnection.prototype_descriptor.getLocalStreams
 	);
-	Object.defineProperty(this, 'addTransceiver', RTCPeerConnection.prototype_descriptor.addTransceiver);
-	Object.defineProperty(this, 'getOrCreateTrack', RTCPeerConnection.prototype_descriptor.getOrCreateTrack);
+	Object.defineProperty(
+		this,
+		'addTransceiver',
+		RTCPeerConnection.prototype_descriptor.addTransceiver
+	);
+	Object.defineProperty(
+		this,
+		'getOrCreateTrack',
+		RTCPeerConnection.prototype_descriptor.getOrCreateTrack
+	);
 
 	// Public atributes.
 	this._localDescription = null;
@@ -306,6 +313,19 @@ RTCPeerConnection.prototype.setRemoteDescription = function (desc) {
 
 	debug('setRemoteDescription() [desc:%o]', desc);
 
+	// Remove extmap-allow-mixed sdp header
+	if (desc && desc.sdp && desc.sdp.indexOf('\na=extmap-allow-mixed') !== -1) {
+		desc = new RTCSessionDescription({
+			type: desc.type,
+			sdp: desc.sdp
+				.split('\n')
+				.filter((line) => {
+					return line.trim() !== 'a=extmap-allow-mixed';
+				})
+				.join('\n')
+		});
+	}
+
 	// "This is no longer necessary, however; RTCPeerConnection.setLocalDescription() and other
 	// methods which take SDP as input now directly accept an object conforming to the RTCSessionDescriptionInit dictionary,
 	// so you don't have to instantiate an RTCSessionDescription yourself.""
@@ -478,6 +498,16 @@ RTCPeerConnection.prototype.addTrack = function (track, ...streams) {
 		this.addStream(stream);
 	}
 
+	var transceiver = new RTCRtpTransceiver(this, {
+		sender: new RTCRtpSender(this, {
+			track: track
+		}),
+		receiver: new RTCRtpReceiver(this),
+		direction: 'sendrecv',
+		currentDirection: null,
+		mid: null
+	});
+
 	for (var streamId in this.localStreams) {
 		if (this.localStreams.hasOwnProperty(streamId)) {
 			// Target provided stream argument or first added stream to group track
@@ -487,6 +517,9 @@ RTCPeerConnection.prototype.addTrack = function (track, ...streams) {
 				exec(null, null, 'iosrtcPlugin', 'RTCPeerConnection_addTrack', [
 					this.pcId,
 					track.id,
+					transceiver._id,
+					transceiver.receiver._id,
+					transceiver.sender._id,
 					streamId
 				]);
 				break;
@@ -496,12 +529,21 @@ RTCPeerConnection.prototype.addTrack = function (track, ...streams) {
 
 	// No Stream matched add track without stream
 	if (!stream) {
-		exec(null, null, 'iosrtcPlugin', 'RTCPeerConnection_addTrack', [this.pcId, track.id, null]);
+		exec(null, null, 'iosrtcPlugin', 'RTCPeerConnection_addTrack', [
+			this.pcId,
+			track.id,
+			transceiver._id,
+			transceiver.receiver._id,
+			transceiver.sender._id,
+			null
+		]);
 	}
 
 	this.getOrCreateTrack(track);
 
-	return new RTCRtpSender(this, { track });
+	this.transceivers.push(transceiver);
+
+	return transceiver.sender;
 };
 
 RTCPeerConnection.prototype.removeTrack = function (sender) {
@@ -527,8 +569,7 @@ RTCPeerConnection.prototype.removeTrack = function (sender) {
 				stream.removeTrack(track);
 				exec(null, null, 'iosrtcPlugin', 'RTCPeerConnection_removeTrack', [
 					this.pcId,
-					track.id,
-					stream.id
+					sender._id
 				]);
 				delete this.tracks[track.id];
 				break;
@@ -543,8 +584,7 @@ RTCPeerConnection.prototype.removeTrack = function (sender) {
 				if (track.id === id) {
 					exec(null, null, 'iosrtcPlugin', 'RTCPeerConnection_removeTrack', [
 						this.pcId,
-						track.id,
-						null
+						sender._id
 					]);
 					delete this.tracks[track.id];
 				}
@@ -554,7 +594,8 @@ RTCPeerConnection.prototype.removeTrack = function (sender) {
 };
 
 RTCPeerConnection.prototype.getOrCreateTrack = function (trackInput) {
-	var { id } = trackInput, existingTrack = this.tracks[id];
+	var { id } = trackInput,
+		existingTrack = this.tracks[id];
 
 	if (existingTrack) {
 		return existingTrack;
@@ -580,14 +621,19 @@ RTCPeerConnection.prototype.addTransceiver = function (trackOrKind, init) {
 		throw new Errors.InvalidStateError('peerconnection is closed');
 	}
 
-	var kind, track = null, self = this, trackIdOrKind;
+	var kind,
+		track = null,
+		self = this,
+		trackIdOrKind;
 	if (trackOrKind instanceof MediaStreamTrack) {
 		kind = trackOrKind.kind;
 		track = trackOrKind;
 		trackIdOrKind = trackOrKind.id;
 	} else {
-		if (!(trackOrKind === "audio" || trackOrKind === "video")) {
-			throw new TypeError("An invalid string was specified as trackOrKind. The string must be either \"audio\" or \"video\".");
+		if (!(trackOrKind === 'audio' || trackOrKind === 'video')) {
+			throw new TypeError(
+				'An invalid string was specified as trackOrKind. The string must be either "audio" or "video".'
+			);
 		}
 		kind = trackOrKind;
 		trackIdOrKind = trackOrKind;
@@ -629,25 +675,34 @@ RTCPeerConnection.prototype.addTransceiver = function (trackOrKind, init) {
 	}
 
 	if (init && init.streams) {
-		initJson.streams = init.streams.map(stream => stream.id);
+		initJson.streams = init.streams.map((stream) => stream.id);
 	} else {
 		initJson.streams = [];
 	}
 
-	debug('addTransceiver() [trackIdOrKind:%s, init:%o, initJson: %o]', trackIdOrKind, init, initJson);
+	debug(
+		'addTransceiver() [trackIdOrKind:%s, init:%o, initJson: %o]',
+		trackIdOrKind,
+		init,
+		initJson
+	);
 
-	addTransceiverToPeerConnection(this, trackIdOrKind, initJson, receiverTrackID)
-		.then(update => {	self.updateTransceiversState(update); })
-		.catch(error => { debugerror("addTransceiver() | failure: %s", error); });
+	addTransceiverToPeerConnection(this, trackIdOrKind, initJson, transceiver)
+		.then((update) => {
+			self.updateTransceiversState(update.transceivers);
+		})
+		.catch((error) => {
+			debugerror('addTransceiver() | failure: %s', error);
+		});
 
 	return transceiver;
 };
 
-RTCPeerConnection.prototype.updateTransceiversState = function(transceivers) {
+RTCPeerConnection.prototype.updateTransceiversState = function (transceivers) {
 	debug('updateTransceiversState()');
 	this.transceivers = transceivers.map((transceiver) => {
 		const existingTransceiver = this.transceivers.find(
-			(localTransceiver) => localTransceiver.receiverId === transceiver.receiver.track.id
+			(localTransceiver) => localTransceiver._id === transceiver.id
 		);
 
 		if (existingTransceiver) {
@@ -766,11 +821,7 @@ RTCPeerConnection.prototype.getStats = function (selector) {
 				return;
 			}
 
-			var res = [];
-			array.forEach(function (stat) {
-				res.push(new RTCStatsReport(stat));
-			});
-			resolve(new RTCStatsResponse(res));
+			resolve(new RTCStatsReport(array));
 		}
 
 		function onResultError(error) {
@@ -903,8 +954,8 @@ function onEvent(data) {
 		case 'track':
 			var track = (event.track = new MediaStreamTrack(data.track));
 			event.receiver = new RTCRtpReceiver(self, { track: track });
-			
-			transceiver = this.transceivers.find(t => t.receiver.track.id === track.id);
+
+			transceiver = this.transceivers.find((t) => t.receiver.track.id === track.id);
 			event.transceiver = transceiver;
 			event.streams = [];