+++ js/RTCRtpTransceiver.js
@@ -9,17 +9,26 @@ module.exports = { RTCRtpTransceiver, addTransceiverToPeerConnection };
 /**
  * Dependencies.
  */
-var
-	debugerror = require('debug')('iosrtc:ERROR:RTCRtpTransceiver'),
+var debugerror = require('debug')('iosrtc:ERROR:RTCRtpTransceiver'),
 	exec = require('cordova/exec'),
+	randomNumber = require('random-number').generator({ min: 10000, max: 99999, integer: true }),
 	EventTarget = require('./EventTarget');
 
 debugerror.log = console.warn.bind(console);
 
-function addTransceiverToPeerConnection(peerConnection, trackIdOrKind, init, receiverTrackId) {
+function addTransceiverToPeerConnection(peerConnection, trackIdOrKind, init, transceiver) {
 	return new Promise((resolve, reject) => {
-		exec(onResultOK, reject, 'iosrtcPlugin', 'RTCPeerConnection_addTransceiver', 
-			[peerConnection.pcId, trackIdOrKind, init, receiverTrackId]);
+		exec(onResultOK, reject, 'iosrtcPlugin', 'RTCPeerConnection_addTransceiver', [
+			peerConnection.pcId,
+			trackIdOrKind,
+			init,
+			transceiver._id,
+			transceiver.sender ? transceiver.sender._id : 0,
+			transceiver.receiver ? transceiver.receiver._id : 0,
+			transceiver.receiver && transceiver.receiver.track
+				? transceiver.receiver.track.id
+				: null
+		]);
 
 		function onResultOK(data) {
 			resolve(data);
@@ -32,9 +41,15 @@ function RTCRtpTransceiver(peerConnection, data) {
 
 	this.peerConnection = peerConnection;
 
-	this._id = data.id;
-	this._receiver = data.receiver instanceof RTCRtpReceiver ? data.receiver : new RTCRtpReceiver(peerConnection, data.receiver);
-	this._sender = data.sender instanceof RTCRtpSender ? data.sender : new RTCRtpSender(peerConnection, data.sender);
+	this._id = data.id || randomNumber();
+	this._receiver =
+		data.receiver instanceof RTCRtpReceiver
+			? data.receiver
+			: new RTCRtpReceiver(peerConnection, data.receiver);
+	this._sender =
+		data.sender instanceof RTCRtpSender
+			? data.sender
+			: new RTCRtpSender(peerConnection, data.sender);
 	this._stopped = false;
 
 	if (data.mid) {
@@ -47,8 +62,8 @@ function RTCRtpTransceiver(peerConnection, data) {
 		this._direction = data.direction;
 		this._currentDirection = data.direction;
 	} else {
-		this._direction = "sendrecv";
-		this._currentDirection = "sendrecv";
+		this._direction = 'sendrecv';
+		this._currentDirection = 'sendrecv';
 	}
 }
 
@@ -69,7 +84,13 @@ Object.defineProperties(RTCRtpTransceiver.prototype, {
 			var self = this;
 			this._direction = direction;
 
-			exec(onResultOK, null, 'iosrtcPlugin', 'RTCPeerConnection_RTCRtpTransceiver_setDirection', [this.peerConnection.pcId, this._id, direction]);
+			exec(
+				onResultOK,
+				null,
+				'iosrtcPlugin',
+				'RTCPeerConnection_RTCRtpTransceiver_setDirection',
+				[this.peerConnection.pcId, this._id, direction]
+			);
 
 			function onResultOK(data) {
 				self.peerConnection.updateTransceiversState(data.transceivers);
@@ -82,22 +103,22 @@ Object.defineProperties(RTCRtpTransceiver.prototype, {
 		}
 	},
 	receiver: {
-		get: function() {
+		get: function () {
 			return this._receiver;
 		}
 	},
 	sender: {
-		get: function() {
+		get: function () {
 			return this._sender;
 		}
 	},
 	stopped: {
-		get: function() {
+		get: function () {
 			return this._stopped;
 		}
 	},
 	receiverId: {
-		get: function() {
+		get: function () {
 			return this._receiver.track.id;
 		}
 	}
@@ -106,7 +127,10 @@ Object.defineProperties(RTCRtpTransceiver.prototype, {
 RTCRtpTransceiver.prototype.stop = function () {
 	var self = this;
 
-	exec(onResultOK, null, 'iosrtcPlugin', 'RTCPeerConnection_RTCRtpTransceiver_stop', [this.peerConnection.pcId, this._id]);
+	exec(onResultOK, null, 'iosrtcPlugin', 'RTCPeerConnection_RTCRtpTransceiver_stop', [
+		this.peerConnection.pcId,
+		this._id
+	]);
 
 	function onResultOK(data) {
 		self.peerConnection.updateTransceiversState(data.transceivers);
@@ -114,8 +138,6 @@ RTCRtpTransceiver.prototype.stop = function () {
 };
 
 RTCRtpTransceiver.prototype.update = function (data) {
-	this._id = data.id;
-
 	if (data.direction) {
 		this._direction = data.direction;
 	}