+++ src/PluginMediaStreamRenderer.swift
@@ -42,19 +42,25 @@ class PluginMediaStreamRenderer : NSObject, RTCEAGLVideoViewDelegate, RTCVideoRe
 			return;
 		}
 
+		let useManualLayoutRenderer = Bundle.main.object(forInfoDictionaryKey: "UseManualLayoutRenderer") as? Bool ?? false
+
+		// The video element view.
+		let guide = webView.safeAreaLayoutGuide;
+		self.elementView = useManualLayoutRenderer
+			? UIView(frame: CGRect(x: 0.0, y: guide.layoutFrame.minY, width: guide.layoutFrame.width, height: guide.layoutFrame.height))
+			: UIView()
+
 		// The effective video view in which the the video stream is shown.
 		// It's placed over the elementView.
 		self.videoView = RTCEAGLVideoView()
 		self.videoView?.isUserInteractionEnabled = false
 
-		// The video element view.
-		let view = UIView()
-		self.elementView = view
-		view.isUserInteractionEnabled = false
-		view.isHidden = true
-		view.backgroundColor = UIColor.black
-		view.addSubview(self.videoView!)
-		view.layer.masksToBounds = true
+		self.elementView.isUserInteractionEnabled = false
+		self.elementView.isHidden = true
+		self.elementView.backgroundColor = UIColor.black
+		self.elementView.addSubview(self.videoView)
+		self.elementView.layer.masksToBounds = true
+		self.elementView.translatesAutoresizingMaskIntoConstraints = false
 
 		// Place the video element view inside the WebView's superview
 		self.webView.addSubview(view)
@@ -64,17 +70,20 @@ class PluginMediaStreamRenderer : NSObject, RTCEAGLVideoViewDelegate, RTCVideoRe
 		// https://stackoverflow.com/questions/46317061/use-safe-area-layout-programmatically
 		// https://developer.apple.com/documentation/uikit/uiview/2891102-safearealayoutguide
 		// https://developer.apple.com/documentation/uikit/
-		if #available(iOS 11.0, *) {
-			let guide = webView.safeAreaLayoutGuide;
-			view.topAnchor.constraint(equalTo: guide.topAnchor).isActive = true
-			view.bottomAnchor.constraint(equalTo: guide.bottomAnchor).isActive = true
-			view.leftAnchor.constraint(equalTo: guide.leftAnchor).isActive = true
-			view.rightAnchor.constraint(equalTo: guide.rightAnchor).isActive = true
-		} else {
-			NSLayoutConstraint(item: view, attribute: .top, relatedBy: .equal, toItem: webView, attribute: .top, multiplier: 1.0, constant: 0).isActive = true
-			NSLayoutConstraint(item: view, attribute: .bottom, relatedBy: .equal, toItem: webView, attribute: .bottom, multiplier: 1.0, constant: 0).isActive = true
-			NSLayoutConstraint(item: view, attribute: .leading, relatedBy: .equal, toItem: webView, attribute: .leading, multiplier: 1.0, constant: 0).isActive = true
-			NSLayoutConstraint(item: view, attribute: .trailing, relatedBy: .equal, toItem: webView, attribute: .trailing, multiplier: 1.0, constant: 0).isActive = true
+
+		let view = self.elementView;
+		if !useManualLayoutRenderer {
+			if #available(iOS 11.0, *) {
+				view.topAnchor.constraint(equalTo: guide.topAnchor).isActive = true
+				view.bottomAnchor.constraint(equalTo: guide.bottomAnchor).isActive = true
+				view.leftAnchor.constraint(equalTo: guide.leftAnchor).isActive = true
+				view.rightAnchor.constraint(equalTo: guide.rightAnchor).isActive = true
+			} else {
+				NSLayoutConstraint(item: view, attribute: .top, relatedBy: .equal, toItem: webView, attribute: .top, multiplier: 1.0, constant: 0).isActive = true
+				NSLayoutConstraint(item: view, attribute: .bottom, relatedBy: .equal, toItem: webView, attribute: .bottom, multiplier: 1.0, constant: 0).isActive = true
+				NSLayoutConstraint(item: view, attribute: .leading, relatedBy: .equal, toItem: webView, attribute: .leading, multiplier: 1.0, constant: 0).isActive = true
+				NSLayoutConstraint(item: view, attribute: .trailing, relatedBy: .equal, toItem: webView, attribute: .trailing, multiplier: 1.0, constant: 0).isActive = true
+			}
 		}
 	}